refactor: 패키지 이동, 네이밍 변경
