+++ android/app/src/main/java/poke/rogue/helper/presentation/type/TypeActivity.kt
@@ -7,7 +7,7 @@ import androidx.activity.viewModels
 import androidx.appcompat.widget.Toolbar
 import poke.rogue.helper.R
 import poke.rogue.helper.databinding.ActivityTypeBinding
-import poke.rogue.helper.presentation.toolbar.ToolbarActivity
+import poke.rogue.helper.presentation.base.toolbar.ToolbarActivity
 import poke.rogue.helper.presentation.type.model.SelectorType
 import poke.rogue.helper.presentation.type.model.TypeUiModel
 import poke.rogue.helper.presentation.type.result.TypeResultAdapter
